Minor updates and Provisioning redone to load Amap
