+++ package.json
@@ -11,6 +11,7 @@
     "@types/enzyme-adapter-react-16": "^1.0.6",
     "@types/webpack-env": "^1.18.0",
     "echarts": "^5.4.0",
+    "echarts-extension-amap": "^1.11.0",
     "echarts-gl": "^2.0.9",
     "echarts-liquidfill": "^3.1.0",
     "echarts-stat": "^1.2.0",
@@ -35,8 +36,5 @@
     "upgrade": "yarn upgrade --latest",
     "watch": "grafana-toolkit plugin:dev --watch"
   },
-  "version": "3.8.0",
-  "dependencies": {
-    "echarts-extension-amap": "^1.11.0"
-  }
+  "version": "3.8.0"
 }
