+++ provisioning/dashboards/maps.json
@@ -52,7 +52,7 @@
           "key": "5f4005ec4d496188a5f6aad42480e83c",
           "plugin": "AMap.Scale,AMap.ToolBar"
         },
-        "getOption": "option = {\n  backgroundColor: 'transparent',\n  // 加载 amap 组件\n  amap: {\n    // 3D模式，无论你使用的是1.x版本还是2.x版本，都建议开启此项以获得更好的渲染体验\n    viewMode: '3D',\n    // 高德地图支持的初始化地图配置\n    // 高德地图初始中心经纬度\n    center: [108.39, 39.9],\n    // 高德地图初始缩放级别\n    zoom: 4,\n    // 是否开启resize\n    resizeEnable: true,\n    // 自定义地图样式主题\n    mapStyle: 'amap://styles/dark',\n    // 移动过程中实时渲染 默认为true 如数据量较大 建议置为false\n    renderOnMoving: true,\n    // ECharts 图层的 zIndex 默认 2000\n    // 从 v1.9.0 起 此配置项已被弃用 请使用 `echartsLayerInteractive` 代替\n    echartsLayerZIndex: 2019,\n    // 设置 ECharts 图层是否可交互 默认为 true\n    // 设置为 false 可实现高德地图自身图层交互\n    // 此配置项从 v1.9.0 起开始支持\n    echartsLayerInteractive: true,\n    // 是否启用大数据模式 默认为 false\n    // 此配置项从 v1.9.0 起开始支持\n    largeMode: false\n    // 说明：如果想要添加卫星、路网等图层\n    // 暂时先不要使用layers配置，因为存在Bug\n    // 建议使用amap.add的方式，使用方式参见最下方代码\n  },\n  series: [\n    {\n      type: 'scatter',\n      // 使用高德地图坐标系\n      coordinateSystem: 'amap',\n      // 数据格式跟在 geo 坐标系上一样，每一项都是 [经度，纬度，数值大小，其它维度...]\n      data: [[120, 30, 8], [120.1, 30.2, 20]],\n      encode: {\n        value: 2\n      }\n    }\n  ]\n};\nreturn option",
+        "getOption": "const amap = {\n  backgroundColor: 'transparent',\n  // 加载 amap 组件\n  amap: {\n    // 3D模式，无论你使用的是1.x版本还是2.x版本，都建议开启此项以获得更好的渲染体验\n    viewMode: '3D',\n    // 高德地图支持的初始化地图配置\n    // 高德地图初始中心经纬度\n    center: [108.39, 39.9],\n    // 高德地图初始缩放级别\n    zoom: 4,\n    // 是否开启resize\n    resizeEnable: true,\n    // 自定义地图样式主题\n    mapStyle: 'amap://styles/dark',\n    // 移动过程中实时渲染 默认为true 如数据量较大 建议置为false\n    renderOnMoving: true,\n    // ECharts 图层的 zIndex 默认 2000\n    // 从 v1.9.0 起 此配置项已被弃用 请使用 `echartsLayerInteractive` 代替\n    echartsLayerZIndex: 2019,\n    // 设置 ECharts 图层是否可交互 默认为 true\n    // 设置为 false 可实现高德地图自身图层交互\n    // 此配置项从 v1.9.0 起开始支持\n    echartsLayerInteractive: true,\n    // 是否启用大数据模式 默认为 false\n    // 此配置项从 v1.9.0 起开始支持\n    largeMode: false\n    // 说明：如果想要添加卫星、路网等图层\n    // 暂时先不要使用layers配置，因为存在Bug\n    // 建议使用amap.add的方式，使用方式参见最下方代码\n  },\n  series: [\n    {\n      type: 'scatter',\n      // 使用高德地图坐标系\n      coordinateSystem: 'amap',\n      // 数据格式跟在 geo 坐标系上一样，每一项都是 [经度，纬度，数值大小，其它维度...]\n      data: [[120, 30, 8], [120.1, 30.2, 20]],\n      encode: {\n        value: 2\n      }\n    }\n  ]\n};\n\n/**\n * Loading\n */\nconst loading = {\n  graphic: {\n    elements: [\n      {\n        type: 'group',\n        left: 'center',\n        top: 'center',\n        children: new Array(7).fill(0).map((val, i) => ({\n          type: 'rect',\n          x: i * 20,\n          shape: {\n            x: 0,\n            y: -40,\n            width: 10,\n            height: 80\n          },\n          style: {\n            fill: '#5470c6'\n          },\n          keyframeAnimation: {\n            duration: 1000,\n            delay: i * 200,\n            loop: true,\n            keyframes: [\n              {\n                percent: 0.5,\n                scaleY: 0.3,\n                easing: 'cubicIn'\n              },\n              {\n                percent: 1,\n                scaleY: 1,\n                easing: 'cubicOut'\n              }\n            ]\n          }\n        }))\n      }\n    ]\n  }\n}\n\n/**\n * Maps are Ready\n */\namapReady = () => {\n  if (!window.AMap) {\n    return setTimeout(() => {\n      amapReady();\n    }, 100)\n  }\n\n  notifySuccess(['Gaode Maps', 'Loaded...']);\n  echartsInstance.setOption(amap, notmerge = true);\n\n  /**\n   * Get AMap extension component\n   */\n  const amapComponent = echartsInstance.getModel().getComponent('amap');\n  const amapInstance = amapComponent.getAMap();\n  amapInstance.addControl(new AMap.Scale());\n  amapInstance.addControl(new AMap.ToolBar());\n\n  /**\n   * Add SatelliteLayer and RoadNetLayer to map\n   */\n  const satelliteLayer = new AMap.TileLayer.Satellite();\n  const roadNetLayer = new AMap.TileLayer.RoadNet();\n  amapInstance.add([satelliteLayer, roadNetLayer]);\n\n  return;\n}\n\nsetTimeout(() => {\n  amapReady();\n}, 100)\n\nreturn loading;",
         "map": "amap",
         "renderer": "canvas"
       },
@@ -138,7 +138,7 @@
     }
   ],
   "refresh": false,
-  "schemaVersion": 36,
+  "schemaVersion": 37,
   "style": "dark",
   "tags": [],
   "templating": {
